pyproject.toml
[project]
name = "leveraged-etf"
version = "0.1.0"
description = "Automated leveraged ETF position-trading system (TQQQ)"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.33.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
include = ["strategy*", "db*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
